Move from directive to service for simplicity
